fix: 사용 중 과목의 CSV 교체 방지 (#221)

+++ src/main/java/edu/handong/csee/histudy/service/CourseService.java
@@ -31,6 +31,9 @@ public class CourseService {
     AcademicTerm currentTerm =
         academicTermRepository.findCurrentSemester().orElseThrow(NoCurrentTermFoundException::new);
     List<Course> courses = toCourses(courseData, currentTerm);
+    if (courseRepository.hasReferences(currentTerm)) {
+      throw new CourseInUseException();
+    }
 
     courseRepository.deleteAllByAcademicTerm(currentTerm);
     courseRepository.saveAll(courses);
